requirements.txt
opencv-python>=4.10.0
SpeechRecognition==3.10.0
openai==1.3.0
numpy==1.24.3
gTTS>=2.3.2
pydub>=0.25.1
python-dotenv>=1.0.1
pyaudio>=0.2.13
tensorflow==2.13.0
tensorflow-macos==2.13.0
keras==2.13.1
typing-extensions==4.5.0
deepface==0.0.79